# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GHF Inventarios API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sociedades:
    get:
      operationId: listSociedades
      tags: [sociedades]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Sociedad" }
    post:
      operationId: createSociedad
      tags: [sociedades]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SociedadInput" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sociedad" }
  /sociedades/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getSociedad
      tags: [sociedades]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sociedad" }
    put:
      operationId: updateSociedad
      tags: [sociedades]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SociedadInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sociedad" }
    delete:
      operationId: deleteSociedad
      tags: [sociedades]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  # ------------------ MARCAS ------------------
...
  /productos:
    get:
      operationId: listProductos
      tags: [productos]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: categoriaId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Producto" }
    post:
      operationId: createProducto
      tags: [productos]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductoInput" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Producto" }
  /productos/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getProducto
      tags: [productos]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Producto" }
    put:
      operationId: updateProducto
      tags: [productos]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductoInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Producto" }
    delete:
      operationId: deleteProducto
      tags: [productos]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  # ------------------ STOCK ------------------
  /stock:
    get:
      operationId: listStock
      tags: [stock]
      parameters:
        - in: query
          name: localId
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StockItem" }
  /stock/movimientos:
    get:
      operationId: listMovimientos
      tags: [stock]
      parameters:
        - in: query
          name: localId
          schema: { type: integer }
        - in: query
          name: limite
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Movimiento" }
    post:
      operationId: createMovimiento
      tags: [stock]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MovimientoInput" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Movimiento" }

  # ------------------ PEDIDOS ------------------
...
  /usuarios:
    get:
      operationId: listUsuarios
      tags: [usuarios]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Usuario" }
    post:
      operationId: createUsuario
      tags: [usuarios]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UsuarioInput" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Usuario" }
  /usuarios/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    put:
      operationId: updateUsuario
      tags: [usuarios]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UsuarioInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Usuario" }
    delete:
      operationId: deleteUsuario
      tags: [usuarios]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    Sociedad:
      type: object
      properties:
        id: { type: integer }
        nombre: { type: string }
        cedulaJuridica: { type: string, nullable: true }
        correo: { type: string, nullable: true }
        telefono: { type: string, nullable: true }
        direccion: { type: string, nullable: true }
        activo: { type: boolean }
        localesCount: { type: integer }
      required: [id, nombre, activo, localesCount]
    SociedadInput:
      type: object
      properties:
        nombre: { type: string }
        cedulaJuridica: { type: string, nullable: true }
        correo: { type: string, nullable: true }
        telefono: { type: string, nullable: true }
        direccion: { type: string, nullable: true }
        activo: { type: boolean }
      required: [nombre]
...
    Producto:
      type: object
      properties:
        id: { type: integer }
        categoriaId: { type: integer, nullable: true }
        categoriaNombre: { type: string, nullable: true }
        nombre: { type: string }
        sku: { type: string, nullable: true }
        descripcion: { type: string, nullable: true }
        precio: { type: number }
        costo: { type: number }
        stockMinimo: { type: integer }
        activo: { type: boolean }
      required: [id, nombre, precio, costo, stockMinimo, activo]
    ProductoInput:
      type: object
      properties:
        categoriaId: { type: integer, nullable: true }
        nombre: { type: string }
        sku: { type: string, nullable: true }
        descripcion: { type: string, nullable: true }
        precio: { type: number }
        costo: { type: number }
        stockMinimo: { type: integer }
        activo: { type: boolean }
      required: [nombre, precio]
    StockItem:
      type: object
      properties:
        productoId: { type: integer }
        productoNombre: { type: string }
        sku: { type: string, nullable: true }
        categoriaNombre: { type: string, nullable: true }
        localId: { type: integer }
        localNombre: { type: string }
        cantidad: { type: integer }
        stockMinimo: { type: integer }
        estado: { type: string, description: "ok | bajo | agotado" }
      required: [productoId, productoNombre, localId, localNombre, cantidad, stockMinimo, estado]
    Movimiento:
      type: object
      properties:
        id: { type: integer }
        productoId: { type: integer }
        productoNombre: { type: string }
        localId: { type: integer }
        localNombre: { type: string }
        tipo: { type: string, description: "entrada | salida | ajuste | venta" }
        cantidad: { type: integer }
        nota: { type: string, nullable: true }
        usuarioNombre: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, productoId, productoNombre, localId, localNombre, tipo, cantidad, createdAt]
    MovimientoInput:
      type: object
      properties:
        productoId: { type: integer }
        localId: { type: integer }
        tipo: { type: string }
        cantidad: { type: integer }
        nota: { type: string, nullable: true }
      required: [productoId, localId, tipo, cantidad]
...
    Usuario:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        username: { type: string }
        email: { type: string, nullable: true }
        role: { type: string }
        localId: { type: integer, nullable: true }
        localNombre: { type: string, nullable: true }
      required: [id, name, username, role]
    UsuarioInput:
      type: object
      properties:
        name: { type: string }
        username: { type: string }
        email: { type: string, nullable: true }
        password: { type: string, nullable: true }
        role: { type: string }
        localId: { type: integer, nullable: true }
      required: [name, username, role]
...