# lib/api-zod/src/generated/types/movimientoInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GHF Inventarios API
 * OpenAPI spec version: 0.1.0
 */

export interface MovimientoInput {
  productoId: number;
  localId: number;
  tipo: string;
  cantidad: number;
  nota?: string | null;
}