# lib/api-zod/src/generated/types/sociedad.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GHF Inventarios API
 * OpenAPI spec version: 0.1.0
 */

export interface Sociedad {
  id: number;
  nombre: string;
  cedulaJuridica?: string | null;
  correo?: string | null;
  telefono?: string | null;
  direccion?: string | null;
  activo: boolean;
  localesCount: number;
}